Code for Quick Featured Image
This script is used to add Quick Featured Image for posts and page listings, You need to add these files to your child theme file, Once files are placed in the child theme call the main quick-featured-image.php file into functions.php, After that you are all set, The features this code provides is you can change the post feature image without having to go to post and edit the image and update the post, you can also remove the feature image on the fly. this code uses Ajax for the features.